+++ src/function/event.ts
@@ -188,10 +188,10 @@ export const tavern_events = {
 export type ListenerType = {
   [iframe_events.MESSAGE_IFRAME_RENDER_STARTED]: (iframe_name: string) => void;
   [iframe_events.MESSAGE_IFRAME_RENDER_ENDED]: (iframe_name: string) => void;
-  [iframe_events.GENERATION_STARTED]: () => void;
-  [iframe_events.STREAM_TOKEN_RECEIVED_FULLY]: (full_text: string) => void;
-  [iframe_events.STREAM_TOKEN_RECEIVED_INCREMENTALLY]: (incremental_text: string) => void;
-  [iframe_events.GENERATION_ENDED]: (text: string) => void;
+  [iframe_events.GENERATION_STARTED]: (generation_id: string) => void;
+  [iframe_events.STREAM_TOKEN_RECEIVED_FULLY]: (full_text: string, generation_id: string) => void;
+  [iframe_events.STREAM_TOKEN_RECEIVED_INCREMENTALLY]: (incremental_text: string, generation_id: string) => void;
+  [iframe_events.GENERATION_ENDED]: (text: string, generation_id: string) => void;
 
   [tavern_events.APP_READY]: () => void;
   [tavern_events.EXTRAS_CONNECTED]: (modules: any) => void;